[Patchset 6] heapprofd client: shared sampler & client lifetime tracking
The primary changes in this patch are:
1) Removing TLS and moving towards a single (shared) allocation sampler.
2) Ensuring that the profiling client is torn down correctly after the
profiling session stops.

The client's lifetime is managed by shared_ptrs, protected by a spinlock. The
spinlock is also reused to mediate access to the now-shared sampler.

General structure of an allocation hook is now:
* call backing implementation (e.g. malloc)
* grab spinlock
* check if there's a valid client in the master shared_ptr (exit hook if not)
* make sampling decision (exit hook if not sampling)
* make an owning copy of the active client shared_ptr
* unlock spinlock
* use the client to record the allocation - heavy work, uses internal locking
  as necessary, shared_ptr copy ensures the client is alive.

Recording frees is similar (but it still uses the timed mutex internally for
now).

Note that we now reject heapprofd_initialize calls if there's already a valid
client (if lazy shutdown has been initated, we do create a new client, even if
the older one is not yet destroyed due to active references).


There are still further changes to make (e.g. lifting the sampler into the
global scope), which will build on this cl.

Bug: 120468745

+++ src/base/circular_queue_unittest.cc
@@ -56,21 +56,21 @@ TEST(CircularQueueTest, Int) {
   ASSERT_EQ(queue.size(), 0u);
   ASSERT_EQ(queue.begin(), queue.end());
 
-  const int kNumInts = 100000;
+  const size_t kNumInts = 100000u;
 
   {
     std::minstd_rand0 rnd_engine(0);
-    for (int i = 0; i < kNumInts; i++) {
+    for (size_t i = 0; i < kNumInts; i++) {
       int n = static_cast<int>(rnd_engine());
       queue.emplace_back(n);
     }
   }
   ASSERT_EQ(queue.size(), kNumInts);
-  ASSERT_EQ(queue.end() - queue.begin(), kNumInts);
+  ASSERT_EQ(static_cast<size_t>(queue.end() - queue.begin()), kNumInts);
   {
     std::minstd_rand0 rnd_engine(0);
     it = queue.begin();
-    for (int i = 0; i < kNumInts; ++i, ++it) {
+    for (size_t i = 0; i < kNumInts; ++i, ++it) {
       ASSERT_LT(it, queue.end());
       ASSERT_EQ(*it, static_cast<int>(rnd_engine()));
     }
